Fixtest&mongodb

+++ backend/src/main/java/de/neuefische/backend/service/BookService.java
@@ -1,20 +1,21 @@
 package de.neuefische.backend.service;
+
 import de.neuefische.backend.model.Book;
+import de.neuefische.backend.model.BookDTO;
 import de.neuefische.backend.repository.BookRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
+import java.util.NoSuchElementException;
 
 @Service
 public class BookService {
 
-    private BookRepository bookRepository;
-    private IDGenerator idGenerator;
+    private final BookRepository bookRepository;
+    private final IDGenerator idGenerator;
 
 
-@Autowired
     public BookService(IDGenerator idGenerator, BookRepository bookRepo){
         this.idGenerator = idGenerator;
         this.bookRepository = bookRepo;
@@ -24,30 +25,37 @@ public class BookService {
         return bookRepository.findAll();
     }
 
-    public Book addBookToList(Book newBook){
+    public Book addBookToList(BookDTO newBook){
         String id = idGenerator.generateID();
-        Book book1 = new Book(id, newBook.getTitle(), newBook.getAuthor(), newBook.getIsbn());
-        return bookRepository.save(book1);
+        Book book1 = new Book(id, newBook.title(), newBook.author(), newBook.isbn());
+        return bookRepository.insert(book1);
     }
     public boolean deleteBook(String id){
-
-       try{
-                bookRepository.deleteById(id);
-                return true;
-            }catch(Exception e){
-                e.getMessage();
-                return false;
-
-            }
-
-   }
-
-
-
-
-    public Optional<Book> getBookByID(String id){
-        return bookRepository.findById(id);
+        if(bookRepository.existsById(id)){
+            bookRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
+    public List<Book> getBookBy(String id, String name, String keyword, String isbn){
+        List<Book> newList = new ArrayList<>();
+        if(id!= null) {
+            newList.add(bookRepository.findById(id).orElseThrow(NoSuchElementException::new));
+        }
+        if(name != null){
+            newList = bookRepository.findBooksByAuthor(name);
+        }
+
+        if(isbn != null){
+            newList.add(bookRepository.findBookByIsbn(isbn));
+        }
+
+        if(keyword != null){
+            newList = bookRepository.findBooksByTitle(keyword);
+        }
+
+        return newList;
+    }
 
 }
